[#93] Add Nebulex.Adapter.Entry and Nebulex.Adapter.Storage behaviours
