+++ lib/nebulex/adapters/replicated.ex
@@ -125,6 +125,8 @@ defmodule Nebulex.Adapters.Replicated do
 
   # Provide Cache Implementation
   @behaviour Nebulex.Adapter
+  @behaviour Nebulex.Adapter.Entry
+  @behaviour Nebulex.Adapter.Storage
   @behaviour Nebulex.Adapter.Queryable
 
   # Inherit default transaction implementation
