+++ lib/nebulex/cache/entry.ex
@@ -223,20 +223,6 @@ defmodule Nebulex.Cache.Entry do
     Adapter.with_meta(name, & &1.touch(&2, key))
   end
 
-  @doc """
-  Implementation for `c:Nebulex.Cache.size/0`.
-  """
-  def size(name) do
-    Adapter.with_meta(name, & &1.size(&2))
-  end
-
-  @doc """
-  Implementation for `c:Nebulex.Cache.flush/0`.
-  """
-  def flush(name) do
-    Adapter.with_meta(name, & &1.flush(&2))
-  end
-
   ## Helpers
 
   defp get_ttl(opts) do
